Lot and Offers

+++ src/main/java/com/carlot/dao/SaleDAO.java
@@ -10,7 +10,7 @@ public interface SaleDAO {
 	public int createOffer (Offer offer)throws BusinessException;
 	
 	// Approve one offer, reject another offers for same car, update car status to sold
-	public int approveOffer (int offerId, int carId)  throws BusinessException;
+	public int approveOffer (long offerId, int carId)  throws BusinessException;
 	public int createLoan (Loan loan)  throws BusinessException;
 	// Create payment and update remaining loan
 	public int createPayment (Payment payment) throws BusinessException;
